+++ src/plonemeeting/restapi/services/add.py
@@ -274,11 +274,18 @@ class ItemPost(BasePost):
 
     @property
     def _optional_fields(self):
-        return self.cfg.listUsedItemAttributes()
+        # XXX manage 'category' manually until
+        # https://support.imio.be/browse/PM-1008 is fixed
+        return self.cfg.listUsedItemAttributes() + ('category', )
 
     @property
     def _active_fields(self):
-        return self.cfg.getUsedItemAttributes()
+        # XXX manage 'category' manually until
+        # https://support.imio.be/browse/PM-1008 is fixed
+        used_item_attrs = self.cfg.getUsedItemAttributes()
+        if not self.cfg.getUseGroupsAsCategories():
+            used_item_attrs += ('category', )
+        return used_item_attrs
 
     def _wf_transition_additional_warning(self, tr):
         warning_message = ""